third_party/blink/web_tests/fast/dom/Document/title-property-set-multiple-times-expected.txt

Test for http://bugs.webkit.org/show_bug.cgi?id=11692. If the title has been explicitly set via document.title, any further <title> tags parsed should not effect the document title.

Expected , got : PASS

Expected First title, got First title: PASS

Expected Second title, got Second title: PASS

Expected Second title, got Second title: PASS

Expected Fourth title, got Fourth title: PASS

Expected Third title, got Third title: PASS

Expected , got : PASS

Expected Fifth title, got Fifth title: PASS
